Move DIRAC.Core.Utilities.Plotting.ObjectLoader.loadObjects to DIRAC.Core.Utilities.ObjectLoader.loadObjects
