+++ src/DIRAC/Core/Utilities/ObjectLoader.py
@@ -6,12 +6,16 @@ from __future__ import print_function
 
 __RCSID__ = "$Id$"
 
-import six
-import re
-import pkgutil
 import collections
+import os
+import re
 
+import pkgutil
+import six
+
+import DIRAC
 from DIRAC import gLogger, S_OK, S_ERROR
+from DIRAC.Core.Utilities import List
 from DIRAC.Core.Utilities import DErrno
 from DIRAC.Core.Utilities import DIRACSingleton
 from DIRAC.Core.Utilities.Extensions import extensionsByPriority, recurseImport
@@ -178,3 +182,57 @@ class ObjectLoader(object):
         modules[modKeyName] = modClass
 
     return S_OK(modules)
+
+
+def loadObjects(path, reFilter=None, parentClass=None):
+  """
+  :param str path: the path to the syetem for example: DIRAC/AccountingSystem
+  :param object reFilter: regular expression used to found the class
+  :param object parentClass: class instance
+  :return: dictionary containing the name of the class and its instance
+  """
+  if not reFilter:
+    reFilter = re.compile(r".*[a-z1-9]\.py$")
+  pathList = List.fromChar(path, "/")
+
+  objectsToLoad = {}
+  # Find which object files match
+  for parentModule in extensionsByPriority():
+    if six.PY3:
+      objDir = os.path.join(os.path.dirname(os.path.dirname(DIRAC.__file__)), parentModule, *pathList)
+    else:
+      objDir = os.path.join(DIRAC.rootPath, parentModule, *pathList)
+    if not os.path.isdir(objDir):
+      continue
+    for objFile in os.listdir(objDir):
+      if reFilter.match(objFile):
+        pythonClassName = objFile[:-3]
+        if pythonClassName not in objectsToLoad:
+          gLogger.info("Adding to load queue %s/%s/%s" % (parentModule, path, pythonClassName))
+          objectsToLoad[pythonClassName] = parentModule
+
+  # Load them!
+  loadedObjects = {}
+
+  for pythonClassName in objectsToLoad:
+    parentModule = objectsToLoad[pythonClassName]
+    try:
+      # Where parentModule can be DIRAC, pathList is something like [ "AccountingSystem", "Client", "Types" ]
+      # And the python class name is.. well, the python class name
+      objPythonPath = "%s.%s.%s" % (parentModule, ".".join(pathList), pythonClassName)
+      objModule = __import__(objPythonPath,
+                             globals(),
+                             locals(), pythonClassName)
+      objClass = getattr(objModule, pythonClassName)
+    except Exception as e:
+      gLogger.error("Can't load type", "%s/%s: %s" % (parentModule, pythonClassName, str(e)))
+      continue
+    if parentClass == objClass:
+      continue
+    if parentClass and not issubclass(objClass, parentClass):
+      gLogger.warn("%s is not a subclass of %s. Skipping" % (objClass, parentClass))
+      continue
+    gLogger.info("Loaded %s" % objPythonPath)
+    loadedObjects[pythonClassName] = objClass
+
+  return loadedObjects
